jt_wazuh_agent_mgr/requirements.txt
tabulate>=0.9.0
rich>=13.0.0
PyYAML>=6.0
requests>=2.28.0
flask>=2.3.0
